move lineage  models to extensions module
