+++ extensions/lineage/src/main/java/org/apache/polaris/extensions/lineage/LineagePersistence.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.polaris.core.lineage;
+package org.apache.polaris.extensions.lineage;
 
 import java.time.Instant;
 import java.util.List;
