+++ gradle/projects.main.properties
@@ -47,6 +47,7 @@ polaris-misc-types=tools/misc-types
 polaris-extensions-federation-hadoop=extensions/federation/hadoop
 polaris-extensions-federation-hive=extensions/federation/hive
 polaris-extensions-federation-bigquery=extensions/federation/bigquery
+polaris-extensions-lineage=extensions/lineage
 polaris-extensions-auth-opa=extensions/auth/opa
 polaris-extensions-auth-ranger=extensions/auth/ranger
 polaris-extensions-semantic-models=extensions/semantic-models
